+++ NEWS.md
@@ -8,6 +8,7 @@
 * Option to disable node drawing with node-by-node control
 * Node-by-node control of node size
 * Aesthetic changes for heatmap and clone-genome distribution plot
+* Add parameters to specify polygon shape and width.
   
 ## Update
 * Fixed angle calculation bug where child angles do not follow